Restructured files and folders. Refactored imports as necessary
