--- outreach-mapping-solutions-fe-react/src/login/login.jsx
@@ -1,3 +1,0 @@
-export function Login() {
-  return <p>Login Portal</p>;
-}
