--- outreach-mapping-solutions-fe-react/src/mapPortal.jsx
@@ -1,3 +1,0 @@
-export function MapPortal() {
-  return <p>Map Portal</p>;
-}
