+++ outreach-mapping-solutions-fe-react/src/portals/login.jsx
@@ -1,0 +1,3 @@
+export default function Login() {
+  return <p>Login Portal</p>;
+}
