+++ outreach-mapping-solutions-fe-react/src/portals/mapPortal.jsx
@@ -1,0 +1,3 @@
+export default function MapPortal() {
+  return <p>Map Portal</p>;
+}
