+++ outreach-mapping-solutions-fe-react/src/portals/reportPortal.jsx
@@ -1,0 +1,3 @@
+export default function ReportPortal() {
+  return <p>Report Portal</p>;
+}
