--- outreach-mapping-solutions-fe-react/src/reportPortal.jsx
@@ -1,3 +1,0 @@
-export function ReportPortal() {
-  return <p>Report Portal</p>;
-}
